requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-celery-beat==2.0.0
django-cors-headers==3.4.0
django-countries==6.1.2
django-filter==2.3.0
django-filters==0.2.1
django-money==1.1
django-rest-framework==0.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
dynaconf==3.0.0
future==0.18.2
hvac==0.10.4
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
loguru==0.5.1
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.4
pathspec==0.8.0
Pillow==7.2.0
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.6.8
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zope.interface==5.1.0
